+++ spec/gherkin/i18n_spec.rb
@@ -184,6 +184,7 @@ module Gherkin
       | eo        | Esperanto           | Esperanto         |
       | es        | Spanish             | español           |
       | et        | Estonian            | eesti keel        |
+      | fa        | Persian             | فارسی             |
       | fi        | Finnish             | suomi             |
       | fr        | French              | français          |
       | he        | Hebrew              | עברית             |