+++ History.md
@@ -1,5 +1,6 @@
 ## [2.11.4](https://github.com/cucumber/gherkin/compare/v2.11.3...v2.11.4)
 
+* [Core] Added Persian/Farsi (Sam Naseri)
 * [JavaScript] Fix lexer instantiation. ([#197](https://github.com/cucumber/gherkin/pull/197) Julien Biezemans)
 
 ## [2.11.3](https://github.com/cucumber/gherkin/compare/v2.11.2...v2.11.3)